Extract source-independent design vocabulary
