+++ crates/celox-sir/Cargo.toml
@@ -9,6 +9,7 @@ edition.workspace     = true
 
 [dependencies]
 celox-analysis = { path = "../celox-analysis" }
+celox-design   = { path = "../celox-design" }
 fxhash         = { workspace = true }
 num-bigint     = { workspace = true }
 num-traits     = { workspace = true }
